[Auth] fix static pages
