+++ src/desktop/apps/auth2/components/AuthStatic.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import { FormSwitcher } from '@artsy/reaction/dist/Components/Authentication/Desktop/FormSwitcher'
 import { ModalType } from '@artsy/reaction/dist/Components/Authentication/Types'
 import { DesktopHeader } from '@artsy/reaction/dist/Components/Authentication/Desktop/Components/DesktopHeader'
+import { handleSubmit } from '../helpers'
 
 interface Props {
   type: string
@@ -17,9 +18,7 @@ export class AuthStatic extends React.Component<Props> {
         <FormSwitcher
           {...this.props}
           type={this.props.type as ModalType}
-          handleSubmit={() => {
-            console.log('here.')
-          }}
+          handleSubmit={handleSubmit.bind(this, this.props.type)}
         />
       </AuthFormContainer>
     )
